Just another commit...
-Robo Pounder (no rage run)
-Crystal Gator (W.I.P)
-Fixed a bunch of small bugs
-Critter cages functional
-Tweaked AI a bunch (not a full AI rework, but more of a small, all
around "general" fix)
-So many new blocks
-ROBO BLOCKS!!
-Fixed void bug
-Made crystal dimension "immersive"
-Gazelle (needs fixes)
-Fixed an annoying rendering bug with some blocks
-Fixed an issue where building the mod into a jar file causes a gradle
duplicate file error
-Figured out how to modify reach, still need to figure out how to modify
it only under a certain condition
-Fixed a GUI rendering bug with the defossilizer
-BiomeType data gen (not complete)
-New advancements
-Fixed a crash with the ultimate fishing rod
-More utility tweaks in IUtilityHelper interface
-Fixed a noclip bug where the robo pounder would send its targets to the
backrooms
-ClientSetupEvent tweaks
-Fixed an auto-enchanting issue
-Fixed some ore configs not working (now all that's left is the damage
number configs)
-Made bone meal work on crystal grass
-Fixed a bug where sometimes ants wouldn't spawn at all
-Tweaked robo mobs generally
-Networking improvements
-Hercules beetle turning into an actual beetle fix
-Stonecutting recipes
-Some more foods
-Fixed even MORE crashes (zamn)
-All around just mostly complete stuff, apart from the AI, defossiliser
working with hoppers, and reach

+++ src/main/java/io/github/chaosawakens/common/items/EnchantedExtendedHitAxeItem.java
@@ -20,7 +20,7 @@ public class EnchantedExtendedHitAxeItem extends ExtendedHitAxeItem implements I
     public AnimationFactory factory = new AnimationFactory(this);
     private final EnchantmentData[] enchantments;
 
-    public EnchantedExtendedHitAxeItem(IItemTier tier, int attackDamageIn, float attackSpeedIn, int reachDistanceIn, int knockBackIn, Properties builderIn, EnchantmentData[] enchantments) {
+    public EnchantedExtendedHitAxeItem(IItemTier tier, int attackDamageIn, float attackSpeedIn, double reachDistanceIn, int knockBackIn, Properties builderIn, EnchantmentData[] enchantments) {
         super(tier, attackDamageIn, attackSpeedIn, reachDistanceIn, knockBackIn, builderIn);
         this.enchantments = enchantments;
     }
